REF Deleting

+++ src/components/FileInput.vue
@@ -11,7 +11,7 @@
       </a-upload>
       <div v-if="gcLogFiles.contents.length > 0 " v-for="content of gcLogFiles.contents">
         <div id="log-file-element-container">
-          {{content.index + 1 }}
+          {{ content.index + 1 }}
           <div id="log-file-element" :style="content.style" @click="activateGCLogfileForSwap(content.index)">
             {{ content.name }}
           </div>
@@ -22,18 +22,16 @@
           <font-awesome-icon v-if="gcLogFiles.actives === 0" id="delete" :icon="['fas', 'arrow-up']"
                              @click="moveGcLogUp(content.index)" />
           <font-awesome-icon v-if="gcLogFiles.actives === 0" id="delete" :icon="['fas', 'trash']"
-                             @click="removeGCLogEntry(content.index)" />
+                             @click="showDeleteGcModal(content.index)" />
         </div>
       </div>
       <a-button v-if="gcLogFiles.contents.length > 0 " class="plot-button" @click="gcLogFiles.createTotalStwChart()">
         Create STW Graph
       </a-button>
-      <a-button v-if="gcLogFiles.contents.length > 0 " class="plot-button" @click="showDeleteAllGcModal">
+      <a-button v-if="gcLogFiles.contents.length > 0 " class="plot-button" @click="showDeleteGcModal(-1)">
         Delete All
       </a-button>
-      <a-modal v-model:visible="deleteModalVisibleGc" title="Basic Modal" @ok="handleOkGc">
-        <p>Delete All Log files from the list?</p>
-      </a-modal>
+      <a-modal v-model:visible="deleteModalVisibleGc" title="Delete?" @ok="handleOkGc"></a-modal>
     </div>
     <div id="request-log">
       <a-upload v-model:file-list="requestsLogList" :multiple="true" name="uploadRequestLog"
@@ -48,7 +46,7 @@
         <div id="log-file-element-container">
           {{ content.index + 1 }}
           <div id="log-file-element" :style="content.style" @click="activateRequestLogFileForSwap(content.index)">
-          {{ content.name }}
+            {{ content.name }}
           </div>
           <font-awesome-icon v-if="content.active && requestLogFiles.actives === 2" id="swap" :icon="['fas', 'repeat']"
                              @click="swapRequests()" />
@@ -57,19 +55,17 @@
           <font-awesome-icon v-if="requestLogFiles.actives === 0" id="delete" :icon="['fas', 'arrow-up']"
                              @click="moveRequestLogUp(content.index)" />
           <font-awesome-icon v-if="requestLogFiles.actives === 0" id="delete" :icon="['fas', 'trash']"
-                             @click="removeRequestLogEntry(content.index)" />
+                             @click="showDeleteReqModal(content.index)" />
         </div>
       </div>
       <a-button v-if="requestLogFiles.contents.length > 0 " class="plot-button"
                 @click="requestLogFiles.createRequestChart()">
         Create Response Graph
       </a-button>
-      <a-button v-if="requestLogFiles.contents.length > 0 " class="plot-button" @click="showDeleteAllReqModal">
+      <a-button v-if="requestLogFiles.contents.length > 0 " class="plot-button" @click="showDeleteReqModal(-1)">
         Delete All
       </a-button>
-      <a-modal v-model:visible="deleteModalVisibleReq" title="Basic Modal" @ok="handleOkReq">
-        <p>Delete All Log files from the list?</p>
-      </a-modal>
+      <a-modal v-model:visible="deleteModalVisibleReq" title="Delete?" @ok="handleOkReq"></a-modal>
     </div>
   </div>
 </template>
@@ -87,23 +83,33 @@ const fileList = ref();
 const requestLogFiles = requestLogStore();
 const requestsLogList = ref();
 
-const deleteModalVisibleReq = ref<boolean>(false)
-const deleteModalVisibleGc = ref<boolean>(false)
+const deleteIndex = ref<number>(-1);
 
-function showDeleteAllReqModal(){
-  deleteModalVisibleReq.value = true
+const deleteModalVisibleReq = ref<boolean>(false);
+const deleteModalVisibleGc = ref<boolean>(false);
+
+function showDeleteReqModal(singleOrAll: number) {
+  deleteIndex.value = singleOrAll;
+  deleteModalVisibleReq.value = true;
 }
-function showDeleteAllGcModal() {
-  deleteModalVisibleGc.value = true
+
+function showDeleteGcModal(singleOrAll: number) {
+  deleteIndex.value = singleOrAll
+  deleteModalVisibleGc.value = true;
 }
+
 function handleOkReq() {
-  requestLogFiles.deleteAll()
-  deleteModalVisibleReq.value = false
+  if (deleteIndex.value === -1)
+    requestLogFiles.deleteAll();
+  else requestLogFiles.removeEntry(deleteIndex.value)
+  deleteModalVisibleReq.value = false;
 }
 
 function handleOkGc() {
-  gcLogFiles.deleteAll()
-  deleteModalVisibleGc.value = false
+  if (deleteIndex.value === -1)
+    gcLogFiles.deleteAll();
+  else gcLogFiles.removeEntry(deleteIndex.value)
+  deleteModalVisibleGc.value = false;
 }
 
 
@@ -131,7 +137,7 @@ function moveGcLogUp(index: number) {
 }
 
 function moveGcDown(index: number) {
-  gcLogFiles.moveDown(index)
+  gcLogFiles.moveDown(index);
 }
 
 
@@ -158,7 +164,7 @@ function moveRequestLogUp(index: number) {
 }
 
 function moveRequestlogDown(index: number) {
-  requestLogFiles.moveDown(index)
+  requestLogFiles.moveDown(index);
 }
